requirements.txt
boto3==1.34.0
fastapi==0.105.0
uvicorn==0.24.0
python-dotenv==1.0.0
pandas==2.1.4
scikit-learn==1.3.2
python-jose==3.3.0
pydantic==2.5.2
schedule==1.2.1
pytest==7.4.3
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
